requirements.txt
httpx>=0.27.0
typer[all]>=0.12.0
python-dotenv>=1.0.0
rich>=13.0.0